feat: rename 'core/banking-institutions' to 'core/bankingInstitutions'
