+++ src/lib/index.ts
@@ -3,6 +3,6 @@ export * from './formatters'
 export * from './transformers'
 export * from './input'
 export * from './sidebar'
-export * from './bank-accounts'
+export * from './accounts'
 export * from './masker'
 export * from './matcher'
